# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bhojnalaya API",
        "description": "REST API documentation for tables, menu items, orders, and order items.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List orders",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Orders fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create an order with optional items",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a single order",
                "operationId": "8114ac185cc1bf301d8862617e497098",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an order and replace its items",
                "operationId": "451beac48796437035c70268bc39ee41",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order",
                "operationId": "983c7813fdcb0d66fd5d1386fa8f6f22",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Table": {
                "required": [
                    "id",
                    "name",
                    "capacity",
                    "status",
                    "is_active"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "T1"
                    },
                    "floor": {
                        "type": "string",
                        "example": "Ground Floor",
                        "nullable": true
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 4
                    },
                    "status": {
                        "type": "string",
                        "example": "blank"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "OrderItem": {
                "required": [
                    "id",
                    "order_id",
                    "menu_item_id",
                    "quantity",
                    "unit_price",
                    "line_total"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "menu_item_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "example": 220
                    },
                    "line_total": {
                        "type": "number",
                        "format": "float",
                        "example": 440
                    },
                    "notes": {
                        "type": "string",
                        "example": "Less spicy",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:15:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:15:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "NestedOrderItemRequest": {
                "required": [
                    "menu_item_id",
                    "quantity"
                ],
                "properties": {
                    "menu_item_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "notes": {
                        "type": "string",
                        "example": "No onion"
                    }
                },
                "type": "object"
            },
            "Order": {
                "required": [
                    "id",
                    "order_number",
                    "order_type",
                    "status",
                    "subtotal",
                    "tax_amount",
                    "total"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "table_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "order_number": {
                        "type": "string",
                        "example": "ORD-1001"
                    },
                    "order_type": {
                        "type": "string",
                        "example": "Dine In"
                    },
                    "status": {
                        "type": "string",
                        "example": "open"
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 440
                    },
                    "tax_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 460
                    },
                    "notes": {
                        "type": "string",
                        "example": "Serve quickly",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:20:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:20:00.000000Z"
                    },
                    "table": {
                        "$ref": "#/components/schemas/Table"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    }
                },
                "type": "object"
            },
            "OrderRequest": {
                "required": [
                    "order_number",
                    "order_type"
                ],
                "properties": {
                    "table_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_number": {
                        "type": "string",
                        "example": "ORD-1001"
                    },
                    "order_type": {
                        "type": "string",
                        "example": "Dine In"
                    },
                    "status": {
                        "type": "string",
                        "example": "open"
                    },
                    "tax_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "notes": {
                        "type": "string",
                        "example": "Less spicy"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NestedOrderItemRequest"
                        }
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The name field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    },
...
}